Refactor feature components: Rename and reorganize files, update imports, and enhance ReusableFeatureCard styling for improved UI consistency
