+++ src/app/features/page.tsx
@@ -1,4 +1,4 @@
-import FeaturePage from "@/components/Features/feature";
+import FeaturePage from "@/components/Features/Feature";
 
 export default function FeatureFrontPage() {
   return <FeaturePage />;
