+++ src/components/Features/ReusableCard.tsx
@@ -13,23 +13,21 @@ export default function ReusableFeatureCard({
   description,
 }: ReusableFeatureCardProps) {
   return (
-    <div className="w-full max-w-sm h-64 flex">
-      <Card className="bg-[var(--custom-card-color)] border border-foreground/10 rounded-xl p-6 sm:p-3 flex-1 flex flex-col sm:hover:shadow-[0_0_24px_0_rgba(34,197,94,0.2)]">
-        <CardHeader className="flex flex-col items-start space-y-4 p-0">
-          <div className="flex items-center justify-center rounded-full bg-chart-5/10 backdrop-blur-md w-12 h-12">
-            {icon &&
-              React.createElement(icon, {
-                className: "text-2xl text-destructive",
-              })}
-          </div>
-          <CardTitle className="font-bold text-xl text-foreground leading-tight">
-            {title}
-          </CardTitle>
-          <CardDescription className="text-muted-foreground text-sm leading-relaxed">
-            {description}
-          </CardDescription>
-        </CardHeader>
-      </Card>
-    </div>
+    <Card className="p-8 hover:scale-[1.03] transition-transform duration-200 min-h-[340px] flex flex-col justify-start w-full max-w-sm mx-auto">
+      <CardHeader className="flex flex-col items-start gap-4 p-0">
+        <div className="flex items-center justify-center rounded-full bg-destructive/20 shadow-lg w-16 h-16 mb-3">
+          {icon &&
+            React.createElement(icon, {
+              className: "text-3xl text-destructive drop-shadow-md",
+            })}
+        </div>
+        <CardTitle className="font-bold text-2xl leading-tight mb-2">
+          {title}
+        </CardTitle>
+        <CardDescription className="text-muted-foreground text-base leading-relaxed">
+          {description}
+        </CardDescription>
+      </CardHeader>
+    </Card>
   );
 }
